agents/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=0.8.0
pydantic>=2.9.0
